removed some redundancies + refactors

+++ src/Common/Security/BitTools.cs
@@ -16,14 +16,14 @@
 
         public static byte RollLeft(byte value, int count)
         {
-            int tmp = (int)value & 0xFF;
+            int tmp = value & 0xFF;
             tmp = tmp << (count % 8);
             return (byte)((tmp & 0xFF) | (tmp >> 8));
         }
 
         public static byte RollRight(byte value, int count)
         {
-            int tmp = (int)value & 0xFF;
+            int tmp = value & 0xFF;
             tmp = (int)((uint)(tmp << 8) >> (count % 8));
 
             return (byte)((tmp & 0xFF) | (int)((uint)tmp >> 8));
